Rebase and remove path_helpers
