+++ lib/berkshelf/mixin/config.rb
@@ -1,5 +1,3 @@
-require 'json'
-
 module Berkshelf
   module Mixin
     module Config
@@ -52,12 +50,38 @@ module Berkshelf
         # @return [String, nil]
         attr_reader :default_location
 
+        # Converts a path to a path usable for your current platform
+        #
+        # @param [String] path
+        #
+        # @return [String]
+        def platform_specific_path(path)
+          if RUBY_PLATFORM =~ /mswin|mingw|windows/
+            system_drive = ENV['SYSTEMDRIVE'] ? ENV['SYSTEMDRIVE'] : ""
+            path         = win_slashify File.join(system_drive, path.split('/')[2..-1])
+          end
+
+          path
+        end
+
         private
           # @abstract
           #   include and override {location} in your class to define the
           #   default location logic
           def location
             default_location || raise(AbstractFunction, "You must implement #{self.class}#location to define default location logic!")
+          end
+
+          # Convert a unixy filepath to a windowsy filepath. Swaps forward slashes for
+          # double backslashes
+          #
+          # @param [String] path
+          #   filepath to convert
+          #
+          # @return [String]
+          #   converted filepath
+          def win_slashify(path)
+            path.gsub(File::SEPARATOR, (File::ALT_SEPARATOR || '\\'))
           end
       end
